requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.33.0
torch>=2.0.0
pandas>=2.0.1
numpy>=1.24.2
prometheus-client>=0.16.0
pytest>=7.4.0
pytest-asyncio>=1.0.0
httpx>=0.24.0
python-jose>=3.3.0
mlflow>=2.4.0
